+++ src/Runner.php
@@ -12,7 +12,7 @@ use Symfony\Component\Console\Output\ConsoleOutput;
 
 class Runner {
 
-    const VERSION = '0.4.1';
+    const VERSION = '0.4.2';
     const ROBOCLASS = 'RoboFile';
     const ROBOFILE = 'RoboFile.php';