+++ tests/unit/CodeceptionTest.php
@@ -20,14 +20,14 @@ class CodeceptionTest extends \Codeception\TestCase\Test
     // tests
     public function testCodeceptionCommand()
     {
-        verify($this->taskCodecept()->getCommand())->equals('vendor/bin/codecept run  ');
+        verify($this->taskCodecept()->getCommand())->equals('vendor/bin/codecept run');
         verify(trim($this->taskCodecept('codecept.phar')->getCommand()))->equals('codecept.phar run');
     }
 
     public function testCodeceptionRun()
     {
         $this->taskCodecept()->run();
-        $this->codecept->verifyInvoked('executeCommand', ['vendor/bin/codecept run  ']);
+        $this->codecept->verifyInvoked('executeCommand', ['vendor/bin/codecept run']);
     }
 
     public function testCodeceptOptions()
@@ -47,10 +47,10 @@ class CodeceptionTest extends \Codeception\TestCase\Test
             ->xml('result.xml')
             ->html()
             ->getCommand()
-        )->equals('codecept run -c ~/Codeception --xml result.xml --html   tests/unit/Codeception');
+        )->equals('codecept run -c ~/Codeception --xml result.xml --html tests/unit/Codeception');
 
-        verify($this->taskCodecept()->debug()->getCommand())->contains(' --debug ');
-        verify($this->taskCodecept()->silent()->getCommand())->contains(' --silent ');
+        verify($this->taskCodecept()->debug()->getCommand())->contains(' --debug');
+        verify($this->taskCodecept()->silent()->getCommand())->contains(' --silent');
         verify($this->taskCodecept()->excludeGroup('g')->getCommand())->contains(' --exclude-group g');
         verify($this->taskCodecept()->tap()->getCommand())->contains('--tap');
         verify($this->taskCodecept()->json()->getCommand())->contains('--json');